Move VisualContainerElement control to WpfMath proj
