+++ WpfMath.Example/MainWindow.xaml
@@ -1,7 +1,7 @@
 ﻿<Window x:Class="WpfMath.Example.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:controls="clr-namespace:WpfMath.Example.Controls"
+        xmlns:controls="clr-namespace:WpfMath.Controls;assembly=WpfMath"
         Title="WPF-Math Example" Width="640" Height="520"
         Loaded="Window_Loaded" Closed="Window_Closed">
     <Window.Resources>
