+++ WpfMath/WpfMath.csproj
@@ -64,6 +64,7 @@
     <Compile Include="CharSymbol.cs" />
     <Compile Include="BigDelimeterAtom.cs" />
     <Compile Include="DelimiterInfo.cs" />
+    <Compile Include="Controls\VisualContainerElement.cs" />
     <Compile Include="Extensions.cs" />
     <Compile Include="PredefinedColorParser.cs" />
     <Compile Include="StyledAtom.cs" />
